+++ release_notes.md
@@ -1,6 +1,8 @@
 ###In Development
+  - [#285](https://github.com/Mehdik/Humanizer/pull/285): Localised ToQuantity and added FarsiQuantifier
   - [#277](https://github.com/MehdiK/Humanizer/pull/277): Added support for custom enum description attribute property names
   - [#276](https://github.com/Mehdik/Humanizer/pull/276): Added Farsi ToOrdinalWords
+
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.26.1...master)